requirements.txt
# Core scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# EEG processing
mne>=1.4.0
pyedflib>=0.1.30
wfdb>=4.1.0

# Deep learning
torch>=2.0.0
torchvision>=0.15.0

# Traditional ML
xgboost>=1.7.0
scikit-learn>=1.2.0
shap>=0.42.0

# Synthetic data generation
ctgan>=0.7.0
sdv>=1.5.0

# API & Backend
fastapi>=0.100.0
uvicorn>=0.22.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Jupyter
jupyterlab>=4.0.0
ipykernel>=6.23.0

# Testing
pytest>=7.3.0
httpx>=0.24.0
